+++ shopify/packages/hybrid-cart/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @brytdesigns/shopify-hybrid-cart
 
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @brytdesigns/shopify-hybrid-cart-api@1.2.12
+
 ## 0.1.13
 
 ### Patch Changes